miscellaneous header tinkering

+++ unsorted_include_todo/zen/particleManager.h
@@ -1,12 +1,14 @@
 #ifndef _ZEN_PARTICLEMANAGER_H
 #define _ZEN_PARTICLEMANAGER_H
 
+#include "types.h"
+
 /*
  * @brief TODO
  */
 namespace zen {
 struct particleManager {
-	void init(unsigned long, unsigned long, unsigned long, float);
+	void init(u32, u32, u32, f32);
 };
 } // namespace zen
